Revert "Refactor Workers"
