+++ Cargo.toml
@@ -10,8 +10,7 @@ members = [
   "./crates/flaky_test",
   "./crates/sb_os",
   "./crates/node",
-  "./crates/cpu_timer",
-  "./crates/event_manager"
+  "./crates/cpu_timer"
 ]
 resolver = "2"
