+++ crates/base/src/lib.rs
@@ -5,7 +5,7 @@ pub mod deno_runtime;
 pub mod errors_rt;
 pub mod js_worker;
 pub mod macros;
-pub mod rt_worker;
 pub mod server;
 pub mod snapshot;
 pub mod utils;
+pub mod worker_ctx;
