--- crates/base/src/rt_worker/implementation/mod.rs
@@ -1,1 +1,0 @@
-pub mod default_handler;
