--- crates/base/src/rt_worker/mod.rs
@@ -1,5 +1,0 @@
-pub mod implementation;
-pub mod utils;
-pub mod worker;
-pub mod worker_ctx;
-pub mod worker_pool;
