+++ crates/base/tests/import_map_tests.rs
@@ -4,7 +4,7 @@ use std::path::Path;
 use tokio::sync::oneshot;
 use urlencoding::encode;
 
-use base::rt_worker::worker_ctx::{create_worker, WorkerRequestMsg};
+use base::worker_ctx::{create_worker, WorkerRequestMsg};
 use sb_worker_context::essentials::{
     UserWorkerRuntimeOpts, WorkerContextInitOpts, WorkerRuntimeOpts,
 };
