+++ crates/base/tests/oak_user_worker_tests.rs
@@ -1,4 +1,4 @@
-use base::rt_worker::worker_ctx::{create_worker, WorkerRequestMsg};
+use base::worker_ctx::{create_worker, WorkerRequestMsg};
 use hyper::{Body, Request, Response};
 use sb_worker_context::essentials::{
     UserWorkerRuntimeOpts, WorkerContextInitOpts, WorkerRuntimeOpts,
