+++ crates/sb_worker_context/Cargo.toml
@@ -17,4 +17,3 @@ hyper.workspace = true
 serde.workspace = true
 tokio.workspace = true
 uuid.workspace = true
-event_manager = { version = "0.1.0", path = "../event_manager" }
