+++ crates/sb_worker_context/events.rs
@@ -1,31 +1,31 @@
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct PseudoEvent {}
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct BootEvent {
     pub boot_time: usize,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct BootFailure {
     pub msg: String,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct UncaughtException {
     pub exception: String,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct LogEvent {
     pub msg: String,
     pub level: LogLevel,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug)]
 pub enum LogLevel {
     Debug,
     Info,
@@ -33,7 +33,7 @@ pub enum LogLevel {
     Error,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug)]
 pub enum WorkerEvents {
     Boot(BootEvent),
     BootFailure(BootFailure),
@@ -51,22 +51,8 @@ pub struct EventMetadata {
     pub execution_id: Option<Uuid>,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct WorkerEventWithMetadata {
     pub event: WorkerEvents,
     pub metadata: EventMetadata,
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-pub enum RawEvent {
-    Event(WorkerEventWithMetadata),
-    Done,
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct IncomingEvent {
-    event_type: Option<String>,
-    data: Option<Vec<u8>>,
-    done: bool,
 }
