+++ crates/sb_worker_context/lib.rs
@@ -1,1 +1,2 @@
 pub mod essentials;
+pub mod events;
