+++ crates/sb_workers/event_worker.js
